# packages/tesslate-marketplace/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Tesslate Federated Marketplace `/v1`
  version: v1
  description: |
    Federated marketplace wire protocol for OpenSail / Tesslate Studio.

    Every response includes `X-Tesslate-Hub-Id` and
    `X-Tesslate-Hub-Api-Version: v1` headers. Endpoints whose capability
    isn't advertised in `/v1/manifest.capabilities[]` return a typed
    `501 unsupported_capability` JSON envelope.
  contact:
    name: Tesslate
    email: marketplace@tesslate.com
servers:
  - url: http://localhost:8800
    description: Local dev
  - url: https://marketplace.tesslate.com
    description: Tesslate Official
...
paths:
...
  /v1/items:
    get:
      tags: [items]
      summary: List items (cache-fed, paginated)
      parameters:
        - in: query
          name: kind
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: sort
          schema:
            type: string
            enum: [featured, newest, popular, name, rating]
      responses:
        "200":
          description: Paginated item list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemList"
        "501": { $ref: "#/components/responses/UnsupportedCapability" }
  /v1/items/{kind}/{slug}:
    get:
      tags: [items]
      summary: Item detail with embedded version list
      parameters:
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Item detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDetail"
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/items/{kind}/{slug}/versions:
    get:
      tags: [items]
      summary: Version list
      parameters:
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemVersion"
  /v1/items/{kind}/{slug}/versions/{version}:
    get:
      tags: [items]
      parameters:
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Slug"
        - $ref: "#/components/parameters/Version"
      responses:
        "200":
          description: Version detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemVersion"
  /v1/items/{kind}/{slug}/versions/{version}/bundle:
    get:
      tags: [items]
      summary: Bundle envelope (signed URL)
      parameters:
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Slug"
        - $ref: "#/components/parameters/Version"
      responses:
        "200":
          description: Bundle envelope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleEnvelope"
        "410":
          description: Version yanked
        "501": { $ref: "#/components/responses/UnsupportedCapability" }
  /v1/items/{kind}/{slug}/versions/{version}/attestation:
    get:
      tags: [items]
      parameters:
        - $ref: "#/components/parameters/Kind"
        - $ref: "#/components/parameters/Slug"
        - $ref: "#/components/parameters/Version"
      responses:
        "200":
          description: Detached signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attestation"
        "501": { $ref: "#/components/responses/UnsupportedCapability" }
...
  /v1/changes:
    get:
      tags: [changes]
      summary: Incremental tombstone-aware sync feed
      parameters:
        - in: query
          name: since
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Changes feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangesFeed"
  /v1/yanks:
    get:
      tags: [yanks]
      summary: Yank-only subset of the changes feed
      parameters:
        - in: query
          name: since
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Yank events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangesFeed"
    post:
      tags: [yanks]
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YankCreate"
      responses:
        "201":
          description: Yank created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Yank"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    Kind:
      in: path
      name: kind
      required: true
      schema:
        type: string
        enum: [agent, skill, mcp_server, base, app, theme, workflow_template]
    Slug:
      in: path
      name: slug
      required: true
      schema: { type: string }
    Version:
      in: path
      name: version
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    UnsupportedCapability:
      description: Capability disabled on this hub
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: unsupported_capability }
              capability: { type: string }
              hub_id: { type: string }
              details: { type: string }
  schemas:
    Pricing:
      type: object
      properties:
        pricing_type:
          type: string
          enum: [free, paid, subscription]
        price_cents: { type: integer }
        currency: { type: string }
        stripe_price_id: { type: string, nullable: true }
        interval: { type: string, nullable: true }
        extras: { type: object, additionalProperties: true, nullable: true }
...
    ItemSummary:
      type: object
      required: [id, kind, slug, name, pricing, updated_at]
      properties:
        id: { type: string }
        kind: { type: string }
        slug: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        category: { type: string, nullable: true }
        icon: { type: string, nullable: true }
        avatar_url: { type: string, nullable: true }
        is_active: { type: boolean }
        is_featured: { type: boolean }
        is_published: { type: boolean }
        pricing: { $ref: "#/components/schemas/Pricing" }
        tags: { type: array, items: { type: string } }
        rating: { type: number }
        reviews_count: { type: integer }
        downloads: { type: integer }
        install_count: { type: integer }
        latest_version: { type: string, nullable: true }
        creator_handle: { type: string, nullable: true }
        updated_at: { type: string, format: date-time }
    ItemDetail:
      allOf:
        - $ref: "#/components/schemas/ItemSummary"
        - type: object
          properties:
            long_description: { type: string, nullable: true }
            features: { type: array, items: { type: string } }
            tech_stack: { type: array, items: { type: string } }
            extra_metadata: { type: object, additionalProperties: true }
            homepage_url: { type: string, nullable: true }
            git_repo_url: { type: string, nullable: true }
            versions:
              type: array
              items: { $ref: "#/components/schemas/ItemVersion" }
    ItemList:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ItemSummary" }
        next_cursor: { type: string, nullable: true }
        has_more: { type: boolean }
        total: { type: integer, nullable: true }
    ItemVersion:
      type: object
      required: [id, version, created_at]
      properties:
        id: { type: string }
        version: { type: string }
        changelog: { type: string, nullable: true }
        is_yanked: { type: boolean }
        yanked_at: { type: string, format: date-time, nullable: true }
        yank_reason: { type: string, nullable: true }
        yank_severity: { type: string, nullable: true }
        is_published: { type: boolean }
        pricing: { $ref: "#/components/schemas/Pricing" }
        manifest: { type: object, additionalProperties: true, nullable: true }
        created_at: { type: string, format: date-time }
    Attestation:
      type: object
      required: [signature, key_id, algorithm]
      properties:
        signature: { type: string }
        key_id: { type: string }
        algorithm: { type: string, default: ed25519 }
    BundleEnvelope:
      type: object
      required: [url, sha256, size_bytes, content_type, archive_format, expires_at]
      properties:
        url: { type: string }
        sha256: { type: string }
        size_bytes: { type: integer }
        content_type: { type: string }
        archive_format: { type: string }
        expires_at: { type: string, format: date-time }
        attestation: { $ref: "#/components/schemas/Attestation" }
...
    ChangeEvent:
      type: object
      required: [op, kind, slug, etag]
      properties:
        op:
          type: string
          enum: [upsert, delete, deactivate, yank, version_remove, pricing_change]
        kind: { type: string }
        slug: { type: string }
        version: { type: string, nullable: true }
        etag: { type: string }
        payload: { type: object, additionalProperties: true, nullable: true }
        created_at: { type: string, format: date-time }
    ChangesFeed:
      type: object
      properties:
        events:
          type: array
          items: { $ref: "#/components/schemas/ChangeEvent" }
        next_etag: { type: string }
        has_more: { type: boolean }
...
    YankCreate:
      type: object
      required: [kind, slug, reason]
      properties:
        kind: { type: string }
        slug: { type: string }
        version: { type: string, nullable: true }
        severity:
          type: string
          enum: [low, medium, critical]
        reason: { type: string }
        requested_by: { type: string, nullable: true }
    Yank:
      type: object
      properties:
        id: { type: string }
        kind: { type: string }
        slug: { type: string }
        version: { type: string, nullable: true }
        severity: { type: string }
        reason: { type: string }
        requested_by: { type: string, nullable: true }
        state: { type: string }
        resolved_at: { type: string, format: date-time, nullable: true }
        resolution: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...